requirements.txt
# Core ML / NLP
torch
transformers
accelerate
trl
peft
datasets

# Data & utilities
pyyaml
tqdm
pandas
numpy
scikit-learn
orjson
